+++ src/ChatMessage.php
@@ -7,10 +7,10 @@ namespace HeroesofAbenez\Chat;
  * ChatMessage
  *
  * @author Jakub Konečný
- * @property-read int $id
- * @property-read string $message
- * @property-read string $when
- * @property-read ChatCharacter $character
+ * @property int $id
+ * @property string $message
+ * @property string $when
+ * @property ChatCharacter $character
  */
 class ChatMessage {
   use \Nette\SmartObject;
@@ -26,21 +26,37 @@ class ChatMessage {
     $this->when = $when;
     $this->character = $character;
   }
-  
+
   public function getId(): int {
     return $this->id;
   }
-  
+
+  protected function setId(int $id): void {
+    $this->id = $id;
+  }
+
   public function getMessage(): string {
     return $this->message;
   }
-  
+
+  protected function setMessage(string $message): void {
+    $this->message = $message;
+  }
+
   public function getWhen(): string {
     return $this->when;
   }
-  
+
+  protected function setWhen(string $when): void {
+    $this->when = $when;
+  }
+
   public function getCharacter(): ChatCharacter {
     return $this->character;
+  }
+
+  protected function setCharacter(ChatCharacter $character): void {
+    $this->character = $character;
   }
 }
 ?>