+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Version 2.1.0-dev
 - raised minimal version of PHP to 7.4
 - used typed properties (possible BC break)
+- made properties of ChatCharacter and ChatMessage writable
 
 Version 2.0.1
 - re-added support for Latte 2.6